feat: add GameObject and interfaces, refactor input system
- Add GameObject class and IGameObject interface
- Move IInputActions into IInput file and remove separate file
- Update VCS configuration for new package
