+++ .idea/.idea.Lunar/.idea/vcs.xml
@@ -5,5 +5,6 @@
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
     <mapping directory="$PROJECT_DIR$/../Packages/Lunar.Adapters.Unity" vcs="Git" />
     <mapping directory="$PROJECT_DIR$/../Packages/Lunar.Adapters.Unity.Addressables" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/../Packages/dev.yarnspinner.unity" vcs="Git" />
   </component>
 </project>
