+++ src/Lunar/Interfaces/IGameObject.cs
@@ -1,0 +1,10 @@
+namespace Lunar.Interfaces
+{
+    /// <summary>
+    /// Interface of game objects in game engine.
+    /// </summary>
+    public interface IGameObject
+    {
+        GameObject GameObject { get; }
+    }
+}
